+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @hyperlane-xyz/cosmos-types
 
+## 13.2.1
+
+## 13.2.0
+
 ## 13.1.1
 
 ### Patch Changes